integrating the changes for reorganisation of code
this includes the integration of changes for reorganisation of code as in the daily closing higher with Heikin-Ashi file in the following commit:

https://github.com/Aprataksh/Munafa/commit/939bbbe7e7b9d303a83ae9d8733e7a0ac1f2499f

Also, this includes an important bugs fix where the invocation of the function for testing if the index is in the same direction has the flag for position past in. If it is a long position it will check if the index is increasing and if it is on a short position, it will check if the index is decreasing.

Unit test: run the file with various parameters and do sanity testing of the results by looking through them to see if they make logical sense. The MACD crossover has yet to be compared with the results on sites like investing.com

This code has been committed by Aayushmaan Rajora.