bug fix:modify the condition to figure out two consecutive green candles
in the condition for determining two consecutive green candles, we want to include those candles for which the open on the current candle is lower than the close on the previous candle. However, the open on the current candle should not be so low that it is even lower than the open of the previous candle. So we want to make sure that the distance between the open of the current candle and the close of the previous candle is less then the distance between the open of the previous candle and the close of the previous candle.

However, we were not testing properly for this condition. We were actually taking the cases where the open of the current candle is either slightly above the open of the previous candle or much lower than the open of the previous candle or much higher than the close of the previous candle.

Unit test: no unit test has been done right now to compare with websites like investing.com and see if the days that are mentioned in the output actually have the candle green and the way we desired.