add the ability to scan for stocks that are closing lower on a daily basis
this code is adapted from the file which scans for stocks that are closing higher on the daily basis.

The output is provided in a CSV files.

Unit test: look at the dates in the output for is equal to 2 for the stock asian paints. Asian paints decreased on June 1, June 4, June 5 June 6, 2018. ensured that there are 3  rows of reflecting thethree days of consecutive reduction in prices. In none of the cases, the high for the day violated the opening price of the previous day. Also looked at the green candles and ensured that none of them were counted by mistake.