integrate the changes of reorganisation of code
integrate the changes for the reorganisation of code which was first done in the file daily closing higher with Heikin-Ashi.

The changes include moving the purchase and sell transactions into a base class and then inheriting all the other classes for daily closing higher with MACD et cetera from the base class.

it also includes changes for providing the ability to take a short position on the stock. However, unlike as in other files, there has been no condition that has been added to test for taking a short position on the stock. This means if we run this copy of the code, it will actually take a short position when a long position is perhaps better.

We will be adding for a short position in later versions of the code.

Unit test: run the code with the following configuration-no trailing stoploss, historical data of 30 minutes, take up a short position with target as 1.5%, stop loss of -5%, maximum days to hold the position as 5 days, maximum number of days to check for daily closing higher as 2.

This code has been submitted by Aayushmaan Rajora