zerodha: Place a stoploss order for the holdings
send a stoploss order for the instruments in the holding that are showing our laws and have crossed a threshold value for the loss.

For example, if the instruments in the holdings have crossed a threshold value of a loss of 3%, the current code will automatically place an order  at a 5% loss.

The values of three and 5% can be modified within the code.

Also take the zerodha specific configuration from the config file.
this code actually prints out the order and the code for making an order placement is commented out
Unit test:execute the code and ensure that the instrument showing lost below 3% are printed out on the console and an order corresponding to 5% is also printed out on the console.