feature: add purchase cut-off time
in order to increase the possibility of intraday trading, we can now add a purchase cut-off time. The intention is to into the trade into the morning so that there is enough time to exit the trade on the same day.
The brokerage charge for a zerodha for intraday is almost seven times lower than the brokerage charge for delivery. This is mainly because of the taxes that are imposed on intraday as compared to delivery.

If we do not want to use a purchase cut-off time, it can be hard coded to our time later than market closing time for example 4 pm

This code has been submitted by Aayushmaan Rajora

Unit test: run the daily closing higher with MACD strategy and visually ensured by looking at the sorted transactions file that there is no purchase transaction before 11 a.m. which is the time that has been configured in the config file.