utilities to fix any problems in downloading data
this commit contains utilities to fix any problems in downloading data.

Specifically, it changes the column header for files do not have date and time as they are column header in the first column. This problem was faced due to a bug in downloading data with the google Finance client API. after the bug has been fixed, we should not be facing this problem again.

The file for a pending historical data is adapted from the file for downloading data it simply appends the historical data files in the directory mentioned in the config file to the downloaded data present in the directory mentioned in the code. The final version of the data files is placed in the directory called final data which is also mentioned in the code.

This code has been submitted by both Aayushmaan Rajora and Aprataksh Anand