determine the beta of the strategy as compared to a corresponding index
determine the beta of the strategy as compared to a corresponding index.. The beta is calculated as follows.

On a daily basis, the total value of the account is computed by adding the value of the holdings with the balance amount. We assume that we start with a balance of Rs. 5 lakh. This can be modified in the file transaction_details.

Then we look at the percentage change on a daily basis in the total value of the account and compare it with the percentage change on a daily basis of the index.

The slope will give us the beta.

Unit test: execute the code in order to calculate the beta for pharmaceutical stocksagainst the pharmaceutical index. Visually check the value of the holdings by comparing it with that in the Ledger and ensure that the stocks are deleted and added properly.

This code has been submitted by Aayushmaan Rajora