fix an error in logging when there is no index data
fix an error in logging when there is no index data. This error appears only when there is no index data to be found and therefore it was difficult to detect it earlier.

Unit test: ensure that the execution proceeds and the logging statements are recorded when there is no index data available.

This code was submitted by Aayushmaan Rajora