Add a config parameter to get the parts to the data with the end of day for indices
add a config parameter to get the path to end of day data for indexes. This parameter is being used in order to calculate the beta of a particular strategy.

Also modify the purchase cut-off time to 5 PM so that the buy signals can be executed throughout the trading day.

For now, this is the default and while executing on optimising a strategy, the purchase cut-off time can be modified as necessary.

Unit test: ensure that the code for calculating the beta of a strategy works properly.

This code has been submitted by Aayushmaan Rajora