scan for NR7/NR4 trades
implement a scanner to look for NR7 and NR4 trades. This is explained  at http://stockcharts.com/school/doku.php?id=chart_school:trading_strategies:narrow_range_day_nr7

in this, we will search for stocks where the difference between the open and close IE the trading range on the eighth day is less then all the trading range is in the preceding seven trading sessions. The parameter for the number of trading sessions to look at can be modified to look at only 4 preceding trading sessions. In such a case, we will be scanning for NR4 trades.

Unit test: look at the output of the scanner and manually verify that AUROPHARMA was indeed satisfying the criteria on April 23, 2018 because it's trading range was 0.9 as opposed to the trading rangein the preceding seven trading sessions