use data of different durations for trade entry and exit
until now, we were using data of the same duration for deciding on trade entry as well as exit. For example, if we used historical data for 30 minutes MACD to determine a trade entry, we will also be selling based on the historical data for 30 minutes.

In a real-life scenario, however, it is possible that once we enter into a position, we will just leave a order to sell the position at the target which will be fulfilled any time we reached the target and need not necessarily be at exactly 30 minute interval (or for that matter, any interval that we are using for our data).

Typically, for trade exit we will be using shorter duration data and therefore we need to use two different files.

In order to abstract the burden of data reading from the strategy file, we have created a new data reader class. In later versions, I may be possible to use the data reader class to read from the database instead of the files. This way the strategy will not be have to modified if the way we store our data changes.

Unit test: run the test using pharmaceuticals stock as well as PSU banks in the Indian market without the changes and then again with the changes for historical 30 minute data. There was no difference in the results.

Then run the file historical 30 minute data for trade entry and historical five-minute data for trade exit. Visually do a sanity check to see that there are no errors-for example, there are no stocks that were left on sold.

This code has been submitted by Aayushmaan Rajora