sscanner for Bollinger band %b
this scanner determines those securities that have been below a particular level of %b  for a specified number of days. The level and the number of days can be specified by the user in the code.

tthe code will output the dates and the ticker symbols that meet the criteria

Unit test: visually compared the values for a stock Aurobindo Pharma with those shown in investing.com. To make it easier, set the %b tto be equal to 0.5 so that we just have to look for values which are below the middle Bollinger band.

This code has met submitted by Aayushmaan Rajora