Provide the support and resistance in separate columns
To make it easier for other processes to read the results of calculation of support and resistance, they need to be placed in separate columns.

The different files for support and resistance have also been merged into one and renamed and therefore this is a new file but it is essentially based on the earlier code for resistance.py. The decision to  whether you support or resistance is taken by looking at the flag that is hardcoded in the source file.

Unit test: run the file and visually conform that the low and high of the range of resistance and support are placed in separate columns

This code has been submitted by Aprataksh Anand