Download the tick data from zerodha
this code will download the tick data from zerodha and put it into an influx database using post requests.

The code will read all the zerodha related information like zerodha API key et cetera from the config. In this version, the user ID for the zerodha needs to be entered in the code.

Unit test:Test to download the data for the entire day.

This code has been submitted by Aprataksh Anand