use commandline parameters to specify the second is as well as the number of days
in order to make the same code work for different durations as well as different number of days and to be invoked through command line to help in automation, we now provide the arguments for the interval of data (duration of candle in other words) and the number of days per wish to download the data from through a command line

Unit test: use this to download the indexes data for five minutes interval. Confirmed that the code will automatically create a directory under the output directory which is picked up from the config. This newly created directory will contain the downloaded as well is the final data. The directory will have the duration of the candle as well as the date on which the data was downloaded as a part of its name.

This code has been submitted by Aprataksh Anand