use command line arguments to specify whether we need to calculate support or resistance
in order to be able to calculate support and resistance on a daily basis using an automated script, we need to provide commandline arguments in order to be able to calculate both resistance and support.

Also, the name of the files for output need to be changed to indicate whether it the file contains the data for resistance all the data for support

Unit test: run the program by providing the argument of 1 for resistance and 0 for support. Confirmed that files with appropriate names have been created in the output directory and also do a sanity check to make sure that the values that are calculated for both support and resistance are different

This code has been submitted by Aprataksh Anand