at the strategy for opening range breakout
the strategy enters a trade when the price of a security crosses the high a low of the opening range which is thee range established for that security at 10 AM in the morning.

In order to increase the probability of monetary gain, we screened the stocks first using the narrow range 7 scanner. If you're taking a long position, we also check if the stock has gone past the previous day's high. We can also check if there is any resistance or support present near the opening range.the stoploss is set at the low of the opening range if you're taking a long position

Right now, the rules for checking the previous day high is valid only for a long position. For short position, we have to check the previous day low which is not added as yet.the stoploss also should be modified to take the high of the opening range if you're taking a short position

Unit test: run this code for nifty 100 and check and validate the trades. We found that Zee Entertainment Ltd does not have historical five-minute data for May 20 is fifth because of which we are taking May 28 data.apart from this problem, validated the transaction then yes bank and they seem to be okay

Also modified the logging so that we are only printing the date. This is because if they include the hours and minutes,  we have to scroll words right if we are looking at the log in a smaller window..

Also modified the narrow range of 7 code to log debug levels for print statement that could executed in almost every loop.  Also modified the name of the output file back to scanner_output.CSV because that becomes more simpler for other strategies to pick up the output of the scanner

The sport has been submitted by Aayushmaan Rajora