modify daily closing trading higher to consider the percentage change in price
the scanner daily closing higher now provides additional data which is the percentage change in the price for which a security has closed consecutively higher.

In order to determine this, it will take the opening price on the first day and the closing price on the nth day, and being the number of days specified by the user

also modify the strategy to enter into a trade only when we have crossed a threshold of percentage change. For example, be will enter a short position only when the percentage change has been 3% or more within two days. Doing this will increase the probability of the stock moving lower

Unit test: validated the percentage change in price in the output is  correct

this quote has been submitted Aprataksh Anand