determine the OHLC from the downloaded data
determine the OHLC from the downloaded data. The data is also written to the CSV file. And it is also written to the database.

Unit test: use the sample database to generate OHLC and look at the CSV file to do a sanity check. The numbers which are in the high column should successively be higher than all the other numbers in the same room for several such stocks.

This code has been submitted by Aprataksh Anand