scan for a crossover of 10 EMA from below
this code will scan for a crossover of 10 EMA from below. It will also output a list of  the entire EMA for every stock in a CSV file. If the price is below the EMA, it will print as a zero in the CSV file. If the price is above the EMA, it will print as a 1 in the CSV file. If there is a crossover, it will print it as 5. It will also display this on the console. If there is no entry, it will print as -5

unit test: execute the scanner with the data from US ETF and visually check that the crossover is being reported correctly

This code has been submitted by Rohit